backend/requirements.txt
fastapi
uvicorn
langchain
faiss-cpu
pymupdf
python-dotenv
sentence-transformers
groq
python-jose
passlib[bcrypt]
pymongo
